requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0
anthropic==0.34.2

